fix(lua): rename vscode lua plugin to vscode-neovim
